cluster_tweets/requirements.txt
mecab-python
twitter
scikit-learn
